# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Exval is a system that allows you to evaluate expressions",
        "title": "Exval API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/evaluate/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expression"
                ],
                "summary": "Evaluate an expression by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expression id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.EvaluateResponse"
                        }
                    }
                }
            }
        },
        "/expressions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expression"
                ],
                "summary": "List all expressions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ExpressionModel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expression"
                ],
                "summary": "Create a new expression",
                "parameters": [
                    {
                        "description": "Expression to create",
                        "name": "expression",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ExpressionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExpressionModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/expressions/{id}": {
            "delete": {
                "tags": [
                    "expression"
                ],
                "summary": "Delete an expression by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expression id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.EvaluateResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "boolean"
                }
            }
        },
        "controllers.ExpressionCreateRequest": {
            "type": "object",
            "required": [
                "expression"
            ],
            "properties": {
                "expression": {
                    "type": "string"
                }
            }
        },
        "models.ExpressionModel": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expression": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}